Refactor golang 'shared models' into 'shared' sub-package
